more warnings cleanups in glui, in particular for type conversions

+++ glui/glui_filebrowser.cpp
@@ -117,7 +117,7 @@ namespace glui {
       list->delete_all();
       if (hFind != INVALID_HANDLE_VALUE) {
         do {
-          int len = strlen(FN.cFileName);
+          int len = (int)strlen(FN.cFileName);
           if (FN.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) {
             item = '\\';
             item += FN.cFileName;
